formta

+++ 0x17-doubly_linked_lists/5-get_dnodeint.c
@@ -17,7 +17,7 @@ dlistint_t *get_dnodeint_at_index(dlistint_t *head, unsigned int index)
 	tmp = head;
 	while (tmp->next != NULL)
 	{
-		if (t == index)
+		if (t == (int)index)
 			return (tmp);
 		tmp = tmp->next;
 		t++;
